MOBILE-4653 quiz: Correctly load offline state in question
