+++ src/core/features/question/components/question/question.ts
@@ -170,6 +170,7 @@ export class CoreQuestionComponent implements OnInit, AsyncDirective {
         // Load local answers if offline is enabled.
         if (this.offlineEnabled && this.component && this.attemptId) {
             await CoreQuestionHelper.loadLocalAnswers(this.question, this.component, this.attemptId);
+            await CoreQuestionHelper.loadLocalQuestionState(this.question,this.attemptId);
         } else {
             this.question.localAnswers = {};
         }
